Move temporal complexity reduction functions to tools
