+++ tests/opf/test_timeseries_reduction.py
@@ -261,85 +261,6 @@ class TestTimeseriesReduction:
         """
         self.edisgo.analyze()
 
-    def test__scored_most_critical_loading_time_interval(self):
-        self.setup_class()
-        self.setup_flexibility_data()
-        self.edisgo.analyze()
-
-        # test with default values
-        ts_crit = timeseries_reduction._scored_most_critical_loading_time_interval(
-            self.edisgo, 24
-        )
-        assert len(ts_crit) == 9
-        assert (
-            ts_crit.loc[0, "time_steps"]
-            == pd.date_range("1/5/2018", periods=24, freq="H")
-        ).all()
-        assert np.isclose(
-            ts_crit.loc[0, "percentage_max_overloaded_components"], 0.96479
-        )
-        assert np.isclose(
-            ts_crit.loc[1, "percentage_max_overloaded_components"], 0.035211
-        )
-
-        # test with non-default values
-        ts_crit = timeseries_reduction._scored_most_critical_loading_time_interval(
-            self.edisgo, 24, time_step_day_start=4, overloading_factor=0.9
-        )
-        assert len(ts_crit) == 9
-        assert (
-            ts_crit.loc[0, "time_steps"]
-            == pd.date_range("1/5/2018 4:00", periods=24, freq="H")
-        ).all()
-        assert ts_crit.loc[0, "percentage_max_overloaded_components"] == 1
-
-    def test__scored_most_critical_voltage_issues_time_interval(self):
-        self.setup_class()
-        self.setup_flexibility_data()
-        self.edisgo.analyze()
-
-        # test with default values
-        ts_crit = (
-            timeseries_reduction._scored_most_critical_voltage_issues_time_interval(
-                self.edisgo, 24
-            )
-        )
-        assert len(ts_crit) == 9
-        assert (
-            ts_crit.loc[0, "time_steps"]
-            == pd.date_range("1/1/2018", periods=24, freq="H")
-        ).all()
-        assert np.isclose(
-            ts_crit.loc[0, "percentage_buses_max_voltage_deviation"], 0.98592
-        )
-        assert np.isclose(ts_crit.loc[1, "percentage_buses_max_voltage_deviation"], 0.0)
-
-        # test with non-default values
-        ts_crit = (
-            timeseries_reduction._scored_most_critical_voltage_issues_time_interval(
-                self.edisgo, 24, time_step_day_start=4, voltage_deviation_factor=0.5
-            )
-        )
-        assert len(ts_crit) == 9
-        assert (
-            ts_crit.loc[0, "time_steps"]
-            == pd.date_range("1/1/2018 4:00", periods=24, freq="H")
-        ).all()
-        assert np.isclose(
-            ts_crit.loc[0, "percentage_buses_max_voltage_deviation"], 0.99296
-        )
-
-    def test_get_most_critical_time_intervals(self):
-        self.setup_class()
-        self.setup_flexibility_data()
-        self.edisgo.analyze()
-        steps = timeseries_reduction.get_most_critical_time_intervals(
-            self.edisgo,
-        )
-
-        assert len(steps) == 3
-        assert len(steps.columns) == 4
-
     def test_distribute_overlying_grid_timeseries(self):
         self.setup_class()
         self.setup_flexibility_data()
